+++ src/mscorlib/shared/System/Collections/ObjectModel/ReadOnlyCollection.cs
