+++ src/mscorlib/shared/System/Runtime/CompilerServices/CustomConstantAttribute.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.CompilerServices
+{
+    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Parameter, Inherited = false)]
+    public abstract class CustomConstantAttribute : Attribute
+    {
+        public abstract Object Value { get; }
+    }
+}
