+++ src/mscorlib/shared/System/Runtime/CompilerServices/MethodImplAttribute.cs
@@ -1,0 +1,29 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.CompilerServices
+{
+    // Custom attribute to specify additional method properties.
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Constructor, Inherited = false)]
+    sealed public class MethodImplAttribute : Attribute
+    {
+        public MethodCodeType MethodCodeType;
+
+        public MethodImplAttribute(MethodImplOptions methodImplOptions)
+        {
+            Value = methodImplOptions;
+        }
+
+        public MethodImplAttribute(short value)
+        {
+            Value = (MethodImplOptions)value;
+        }
+
+        public MethodImplAttribute()
+        {
+        }
+
+        public MethodImplOptions Value { get; }
+    }
+}
