+++ src/mscorlib/shared/System/Runtime/InteropServices/FieldOffsetAttribute.cs
@@ -1,0 +1,19 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Field, Inherited = false)]
+    public sealed class FieldOffsetAttribute : Attribute
+    {
+        private int _val;
+
+        public FieldOffsetAttribute(int offset)
+        {
+            _val = offset;
+        }
+
+        public int Value => _val;
+    }
+}
