+++ src/mscorlib/shared/System/Runtime/InteropServices/OutAttribute.cs
@@ -1,0 +1,14 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
+    public sealed class OutAttribute : Attribute
+    {
+        public OutAttribute()
+        {
+        }
+    }
+}
