+++ src/mscorlib/shared/System/Runtime/InteropServices/StructLayoutAttribute.cs
@@ -1,0 +1,28 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.InteropServices
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, Inherited = false)]
+    public sealed class StructLayoutAttribute : Attribute
+    {
+        private LayoutKind _val;
+
+        public StructLayoutAttribute(LayoutKind layoutKind)
+        {
+            _val = layoutKind;
+        }
+
+        public StructLayoutAttribute(short layoutKind)
+        {
+            _val = (LayoutKind)layoutKind;
+        }
+
+        public LayoutKind Value => _val;
+
+        public int Pack;
+        public int Size;
+        public CharSet CharSet;
+    }
+}
